requirements.txt
esper
numpy
pillow
pyffmpeg
requests
tqdm
moderngl
msgspec
yaspin